Complete program: This is a solution to the "Find Angle MBC" challenge for Python on hackerrank.com. The original problem can be found here:
https://www.hackerrank.com/challenges/find-angle

The goal is to calculate the angle MBC of a right triangle ABC whose hypotenuse AC has midpoint M, creating sub-triangle MBC. It is done through the use of math functions. Two lines of input, separated by a newline, can be entered, being the length (in real numbers) of side AB and side BC, respectively. The angle MBC is then printed in degrees rounded to the nearest whole int.